% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot.R
\name{chatbot_piece_path}
\alias{chatbot_piece_path}
\title{Returns the file path for a chatbot piece}
\usage{
chatbot_piece_path(name, piece, root_dir = Sys.getenv("CHATBOT_ROOT"))
}
\arguments{
\item{name}{name of the chatbot}

\item{piece}{which piece of the systme prompt--persona, instructions, context, protocol--to get}

\item{root_dir}{root directory of chatbots}
}
\description{
Returns the file path for a chatbot piece
}
