% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persona.R
\name{chaz_persona}
\alias{chaz_persona}
\title{Returns a list of `description` and `prompt` and optionally creates a new
 larkdown file.}
\usage{
chaz_persona(
  input,
  file = NULL,
  open = !is.null(file),
  endpoint_url = getOption("chaz_endpoint_url", default = "")
)
}
\arguments{
\item{input}{Request for a persona to generate}

\item{file}{if not NULL, create}
}
\description{
Returns a list of `description` and `prompt` and optionally creates a new
 larkdown file.
}
