% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{convert_image_to_bytestring}
\alias{convert_image_to_bytestring}
\title{Returns a bytestring of an image, rescaled to specified width}
\usage{
convert_image_to_bytestring(path, width = 200, show_image = FALSE)
}
\arguments{
\item{path}{passed to `magick::image_read()`}

\item{width}{resized image width in pixels. Resize will happen before encoding.}
}
\description{
Returns a bytestring of an image, rescaled to specified width
}
