% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{count_tokens}
\alias{count_tokens}
\title{Return token count and pricing for gpt-4 and gpt-3.5}
\usage{
count_tokens(input, encoding = "cl100k_base")
}
\arguments{
\item{input}{character string to pass as llm input.}

\item{encoding}{passed to `tiktoken.get_encoding`}
}
\description{
Return token count and pricing for gpt-4 and gpt-3.5
}
