% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot.R
\name{create_path}
\alias{create_path}
\title{Helper function to create files and directories as needed}
\usage{
create_path(base_dir, path, edit = FALSE)
}
\arguments{
\item{edit}{if TRUE, open files for editing}
}
\description{
Helper function to create files and directories as needed
}
