% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{current_document}
\alias{current_document}
\title{Returns the full path of the current document, saving it first.}
\usage{
current_document()
}
\description{
Returns the full path of the current document, saving it first.
}
