% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-gen.R
\name{dalle_image}
\alias{dalle_image}
\title{Create an image using Dall-E}
\usage{
dalle_image(
  prompt,
  model = "dall-e-3",
  size = "1024x1024",
  quality = "standard",
  n = 1L
)
}
\arguments{
\item{prompt}{the image-generation prompt}

\item{model}{model name}

\item{size}{resolution of image}

\item{quality}{quality of image}

\item{n}{Number of images to generate}
}
\description{
Create an image using Dall-E
}
