% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{dir_str}
\alias{dir_str}
\title{Return an ASCII representation of directory structure}
\usage{
dir_str(dir, level = 0, pattern = NULL)
}
\arguments{
\item{dir}{path to a directory}

\item{level}{used in recursion}

\item{pattern}{passed to list.files()}
}
\description{
Return an ASCII representation of directory structure
}
