% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot.R
\name{edit_chatbot_piece}
\alias{edit_chatbot_piece}
\title{Opens the chatbot piece file in rstudio}
\usage{
edit_chatbot_piece(name, piece, root_dir = Sys.getenv("CHATBOT_ROOT"))
}
\arguments{
\item{name}{name of the chatbot}

\item{piece}{which piece of the systme prompt--persona, instructions, context, protocol--to get}

\item{root_dir}{root directory of chatbots}
}
\description{
Opens the chatbot piece file in rstudio
}
