% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_date}
\alias{format_date}
\title{Utility function to format a date}
\usage{
format_date(date)
}
\arguments{
\item{date}{The date to format}
}
\value{
The formatted date
}
\description{
Utility function to format a date
}
