% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_random_number}
\alias{generate_random_number}
\title{Utility function to generate a random number}
\usage{
generate_random_number()
}
\value{
A random number between 0 and 1
}
\description{
Utility function to generate a random number
}
