% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_larkdown.R
\name{get_system_message_text}
\alias{get_system_message_text}
\title{Returns a character string with the text of a message lists's system message}
\usage{
get_system_message_text(message_list)
}
\arguments{
\item{message_list}{a list of LangChain `BaseMessage` objects}
}
\description{
Returns a character string with the text of a message lists's system message
}
