% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{image_context}
\alias{image_context}
\title{For use with Larkdown, produces a bytestring when knitted such that `larkdown.parse_larkdown()` creates image-augmented messages}
\usage{
image_context(path, width = 200, show_image = interactive())
}
\arguments{
\item{path}{passed to `magick::image_read()`}

\item{width}{resized image width in pixels. Resize will happen before encoding.}
}
\description{
For use with Larkdown, produces a bytestring when knitted such that `larkdown.parse_larkdown()` creates image-augmented messages
}
