% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_larkdown}
\alias{import_larkdown}
\title{Wrapper for `reticulate::import_from_path()` to pull in larkdown module}
\usage{
import_larkdown()
}
\description{
Wrapper for `reticulate::import_from_path()` to pull in larkdown module
}
