% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{jupyter_text}
\alias{jupyter_text}
\title{Read the text of a jupyter notebook to a formatted string}
\usage{
jupyter_text(path)
}
\arguments{
\item{path}{the path to the jupyter notebook}
}
\description{
Read the text of a jupyter notebook to a formatted string
}
