% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{knit_and_stream_current}
\alias{knit_and_stream_current}
\title{knit the current document, parse to messages, and stream from endpoint}
\usage{
knit_and_stream_current(
  endpoint_url = getOption("ld_endpoint_url"),
  prompt = getOption("ld_prompt", default = "@")
)
}
\description{
knit the current document, parse to messages, and stream from endpoint
}
