% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{knit_to_messages}
\alias{knit_to_messages}
\title{Returns the result of calling `lardkown.parse_larkdown()` on the specified file}
\usage{
knit_to_messages(file = current_document())
}
\arguments{
\item{file}{defaults to `current_document()`}
}
\description{
Returns the result of calling `lardkown.parse_larkdown()` on the specified file
}
