% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_larkdown.R
\name{ld_convo_text}
\alias{ld_convo_text}
\title{Creates a larkdown-formatted character string from a list of messages}
\usage{
ld_convo_text(message_list, larkdown_prompt = "@")
}
\arguments{
\item{message_list}{a list of LangChain `BaseMessage` objects}

\item{larkdown_prompt}{Larkdown prompt character to use, defaults to `"@"`}
}
\description{
Used to populate a larkdown file (which is just a textfile) from a message list
}
