% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_larkdown.R
\name{ld_message_type}
\alias{ld_message_type}
\title{Returns "system", "human", or "ai"}
\usage{
ld_message_type(message)
}
\arguments{
\item{message}{LangChain `BaseMessage` object}
}
\description{
Returns "system", "human", or "ai"
}
