% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ld_register_endpoints}
\alias{ld_register_endpoints}
\title{Sets larkdown endpoint url options.}
\usage{
ld_register_endpoints(root_url)
}
\arguments{
\item{root_url}{url of larkdown langserve project root}
}
\description{
Sets larkdown endpoint url options.
}
