% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rag.R
\name{ld_search}
\alias{ld_search}
\title{RAG-over-larkdown function}
\usage{
ld_search(
  input,
  endpoint_url = getOption("ld_search_endpoint_url", default = "")
)
}
\arguments{
\item{input}{text input for searching}

\item{endpoint_url}{best to set this via `ld_register_endpoints()`.}
}
\description{
Relies on a deployed langserve endpoint
}
