% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{ld_stream_current}
\alias{ld_stream_current}
\title{Parse the current document and stream response back to it}
\usage{
ld_stream_current()
}
\description{
Parse the current document and stream response back to it
}
