% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot.R
\name{list_chatbots}
\alias{list_chatbots}
\title{List all currently available chatbots}
\usage{
list_chatbots(root_dir = Sys.getenv("CHATBOT_ROOT"))
}
\arguments{
\item{root_dir}{root directory of chatbots}
}
\description{
List all currently available chatbots
}
