% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{messages_text}
\alias{messages_text}
\title{Returns a character vector (length 1) of concatenated message text}
\usage{
messages_text(message_list)
}
\arguments{
\item{message_list}{list of BaseMessage objects}
}
\description{
Returns a character vector (length 1) of concatenated message text
}
