% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot.R
\name{new_chatbot}
\alias{new_chatbot}
\title{Creates a new chatbot config directory}
\usage{
new_chatbot(name, edit = TRUE)
}
\arguments{
\item{name}{name of the chatbot}

\item{edit}{if TRUE, open files for editing}
}
\description{
Creates a new chatbot config directory
}
