% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_larkdown}
\alias{new_larkdown}
\title{Creates a new Rmarkdown file with a Larkdown template, then opens the file}
\usage{
new_larkdown(
  filename = ts_filename(prefix = "larkdown"),
  template = "larkdown-file",
  package = "larkdown"
)
}
\description{
Creates a new Rmarkdown file with a Larkdown template, then opens the file
}
