% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot.R
\name{read_chatbot_piece}
\alias{read_chatbot_piece}
\title{Returns the file text for a component of a chatbot's system prompt.}
\usage{
read_chatbot_piece(chatbot_name, piece, root_dir = Sys.getenv("CHATBOT_ROOT"))
}
\arguments{
\item{chatbot_name}{name of the chatbot}

\item{piece}{which piece of the systme prompt--persona, instructions, context, protocol--to get}

\item{root_dir}{root directory of chatbots}
}
\description{
Returns the file text for a component of a chatbot's system prompt.
}
