% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tavily.R
\name{tavily_search}
\alias{tavily_search}
\title{Wrapper around langchain TavilySearchResults}
\usage{
tavily_search(
  text,
  max_results = 5,
  search_depth = "advanced",
  include_answer = TRUE,
  include_raw_content = TRUE,
  include_images = TRUE
)
}
\arguments{
\item{text}{search string}

\item{max_results, search_depth, include_anser, include_raw_content, include_raw_images}{passed to TavilySearchResults}
}
\description{
Returns a list of search results
}
