% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ts_filename}
\alias{ts_filename}
\title{Timestamp filename creator}
\usage{
ts_filename(prefix = "", suffix = "", ext = ".Rmd")
}
\arguments{
\item{prefix, suffix, ext}{appended to the timestamp to create the filename}
}
\description{
Timestamp filename creator
}
