% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{yt_desc}
\alias{yt_desc}
\title{Returns the description text from a youtube video}
\usage{
yt_desc(yt_url)
}
\arguments{
\item{yt_url}{url of youtube video}
}
\description{
Works around a currently open issue with pytube package: https://github.com/pytube/pytube/issues/1626#issuecomment-1775501965
}
