% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatbot.R
\name{chatbot_sys_prompt}
\alias{chatbot_sys_prompt}
\title{Returns the system prompt for a chatbot}
\usage{
chatbot_sys_prompt(chatbot_name, root_dir = Sys.getenv("CHATBOT_ROOT"))
}
\arguments{
\item{chatbot_name}{name of the chatbot}

\item{root_dir}{root directory of chatbots}
}
\description{
To be used with `larkdown::new_larkdown()`
}
