% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{dir_text}
\alias{dir_text}
\title{Return text of files for all files in a directory}
\usage{
dir_text(path, pattern = ".*")
}
\arguments{
\item{path}{Path to directory}
}
\description{
Return text of files for all files in a directory
}
