% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{journal}
\alias{journal}
\title{Launches a journaling session for the current day.}
\usage{
journal(base_path = Sys.getenv("LARKDOWN_DIR"))
}
\description{
Launches a journaling session for the current day.
}
