% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_larkdown.R
\name{journal}
\alias{journal}
\title{Launches a journaling session for the current day.}
\usage{
journal(base_path = Sys.getenv("LARKDOWN_DIR"))
}
\arguments{
\item{base_path}{directory in which to save the journal}
}
\description{
Launches a journaling session for the current day.
}
