% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_larkdown.R
\name{new_larkdown}
\alias{new_larkdown}
\title{Creates a new Rmarkdown file with a Larkdown template, then opens the file}
\usage{
new_larkdown(
  filename = ts_filename(prefix = "larkdown"),
  system_message = NULL,
  message_list = list(),
  header_text = "",
  template = "larkdown-glue",
  open = TRUE
)
}
\description{
Creates a new Rmarkdown file with a Larkdown template, then opens the file
}
