% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{stream_current}
\alias{stream_current}
\title{Parse the current document and stream response back to it}
\usage{
stream_current(
  endpoint_url = getOption("ld_endpoint_url"),
  prompt = getOption("ld_prompt", default = "@")
)
}
\arguments{
\item{endpoint_url}{endpoint of LangServe endpoint}

\item{prompt}{Larkdown prompt character}
}
\description{
Parse the current document and stream response back to it
}
